requirements.txt
boto3~=1.14.5
checkdmarc~=4.2.4
pytest~=5.4.3
dnspython~=2.1.0
